+++ src/event-listeners/KeyDownEventListener.test.ts
@@ -1,17 +1,23 @@
-import { beforeEach, describe, expect, it, vitest } from 'vitest'
+import { beforeEach, describe, expect, it, SpyInstance, vitest } from 'vitest'
 import UserActionHandler from '../user-action/UserActionHandler'
 import { fireEvent, getAllByRole, getByRole, waitFor } from '@testing-library/dom'
 import { nop } from '../utils/nop'
 import createElementInJSDOM from '../test-utils/createElementInJSDOM'
 import KeyDownEventListener from '../event-listeners/KeyDownEventListener'
+import UserActionHistory from '../user-actions-history/UserActionHistory'
+import MemoryUserActionsHistory from '../user-actions-history/MemoryUserActionsHistory'
 
 describe('KeyDownEventListener', () => {
   describe('listener', () => {
-    const userActionHandler = new UserActionHandler('aaa-111', 0, nop)
-    const runSpy = vitest.spyOn(userActionHandler, 'handle')
+    let userActionHistory: UserActionHistory
+    let userActionHandler: UserActionHandler
+    let handleSpy: SpyInstance
 
     beforeEach(() => {
       vitest.restoreAllMocks()
+      userActionHistory = new MemoryUserActionsHistory()
+      userActionHandler = new UserActionHandler('aaa-111', 0, nop, userActionHistory)
+      handleSpy = vitest.spyOn(userActionHandler, 'handle')
     })
 
     it('calls listener when key down event been fired', async () => {
@@ -23,13 +29,13 @@ describe('KeyDownEventListener', () => {
         'div',
       )
 
-      new KeyDownEventListener(userActionHandler, domWindow).init()
+      new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
       const button = await waitFor(() => getByRole(element, 'checkbox'))
 
       fireEvent.keyDown(button)
 
       await waitFor(() => {
-        expect(runSpy).toHaveBeenCalledOnce()
+        expect(handleSpy).toHaveBeenCalledOnce()
       })
     })
 
@@ -42,13 +48,34 @@ describe('KeyDownEventListener', () => {
         'div',
       )
 
-      new KeyDownEventListener(userActionHandler, domWindow).init()
+      new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
       const div = await waitFor(() => getByRole(element, 'cell'))
 
       fireEvent.keyDown(div)
 
       await waitFor(() => {
-        expect(runSpy).toHaveBeenCalledOnce()
+        expect(handleSpy).toHaveBeenCalledOnce()
+      })
+    })
+
+    it('does not call listener multiple time when key down event is fired consecutively', async () => {
+      const { element, domWindow } = createElementInJSDOM(
+        `
+                <div>
+                    <div role="cell"/>
+                </div>`,
+        'div',
+      )
+
+      new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
+      const div = await waitFor(() => getByRole(element, 'cell'))
+
+      for (let i = 0; i < 10; i++) fireEvent.keyDown(div)
+
+      await waitFor(() => {}, { timeout: 200 })
+
+      await waitFor(() => {
+        expect(handleSpy).toHaveBeenCalledOnce()
       })
     })
 
@@ -63,7 +90,7 @@ describe('KeyDownEventListener', () => {
         'div',
       )
 
-      new KeyDownEventListener(userActionHandler, domWindow).init()
+      new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
 
       const inputs: HTMLElement[] = await waitFor(() => getAllByRole(element, 'textbox'))
       const search = await waitFor(() => getByRole(element, 'searchbox'))
@@ -76,7 +103,7 @@ describe('KeyDownEventListener', () => {
       await waitFor(() => {}, { timeout: 500 })
 
       await waitFor(() => {
-        expect(runSpy).toHaveBeenCalledTimes(0)
+        expect(handleSpy).toHaveBeenCalledTimes(0)
       })
     })
 
@@ -91,7 +118,7 @@ describe('KeyDownEventListener', () => {
         'div',
       )
 
-      new KeyDownEventListener(userActionHandler, domWindow).init()
+      new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
 
       const inputs: HTMLElement[] = await waitFor(() => getAllByRole(element, 'textbox'))
       const search = await waitFor(() => getByRole(element, 'searchbox'))
@@ -102,7 +129,7 @@ describe('KeyDownEventListener', () => {
       fireEvent.keyDown(search, { code: 'Tab' })
 
       await waitFor(() => {
-        expect(runSpy).toHaveBeenCalledTimes(3)
+        expect(handleSpy).toHaveBeenCalledTimes(3)
       })
     })
   })