+++ src/user-actions-history/UserActionHistory.ts
@@ -1,0 +1,9 @@
+import { UserAction } from '../types'
+
+export default interface UserActionHistory {
+  push: (userAction: UserAction) => void
+
+  getLast: () => UserAction
+
+  getUserActionsHistory: () => UserAction[]
+}